Some tweaks for the multi element reader stuff
