+++ src/main/scala/fm/xml/XmlReader.scala
@@ -21,19 +21,17 @@ import fm.common.{Logging, Resource}
 import fm.lazyseq.ResourceLazySeq
 import fm.xml.RichXMLStreamReader2.toRichXMLStreamReader2
 import java.io.{InputStream, Reader, StringReader}
-import javax.xml.bind.{JAXBContext, Unmarshaller}
 import javax.xml.stream.XMLInputFactory
-import javax.xml.stream.XMLStreamConstants.START_ELEMENT
 import org.codehaus.stax2.XMLStreamReader2
 import org.codehaus.stax2.util.StreamReader2Delegate
 import scala.collection.mutable
-import scala.reflect.{ClassTag, classTag}
+import scala.reflect.ClassTag
 import scala.util.Try
 
 object XmlReader {
   // Backward source-compatible support for specifying a simple itemPath/element type
   def apply[T: ClassTag](rootName: String, itemPath: String, defaultNamespaceURI: String, overrideDefaultNamespaceURI: String, resource: Resource[Reader]): XmlReader[T] = {
-    val path: XmlReaderPath[T, T] = XmlReaderPath[T, T](itemPath, identity(_))
+    val path: XmlReaderPath[T, T] = XmlReaderPath[T, T](itemPath, identity)
 
     new XmlReader[T](
       rootName,
@@ -44,17 +42,31 @@ object XmlReader {
     )
   }
 
+  def apply[T](rootName: String, resource: Resource[Reader], targetHead: XmlReaderPath[_,T], targetsRest: XmlReaderPath[_,T]*): XmlReader[T] = {
+    apply(rootName, resource, (targetHead +: targetsRest).toIndexedSeq)
+  }
+
+  def apply[T](rootName: String, resource: Resource[Reader], targets: IndexedSeq[XmlReaderPath[_,T]]): XmlReader[T] = {
+    XmlReader(
+      rootName = rootName,
+      defaultNamespaceURI = "",
+      overrideDefaultNamespaceURI = "",
+      resource = resource,
+      targets = targets
+    )
+  }
+
   private val inputFactory: WstxInputFactory = new WstxInputFactory()
   inputFactory.setProperty(XMLInputFactory.SUPPORT_DTD, false)
   inputFactory.configureForSpeed()
 
-  import Resource._
-
   def withXMLStreamReader2[T](s: String)(f: XMLStreamReader2 => T): T = {
+    import Resource._
     Resource.using(inputFactory.createXMLStreamReader(new StringReader(s)).asInstanceOf[XMLStreamReader2])(f)
   }
 
   def withXMLStreamReader2[T](is: InputStream)(f: XMLStreamReader2 => T): T = {
+    import Resource._
     Resource.using(inputFactory.createXMLStreamReader(is).asInstanceOf[XMLStreamReader2])(f)
   }
 
@@ -106,84 +118,11 @@ object XmlReader {
   }
 }
 
-
-object XmlReaderPath {
-  // Apply helper for reading a non-mapped value
-  def apply[XmlValue: ClassTag](itemPath: String): XmlReaderPath[XmlValue, XmlValue] = XmlReaderPath[XmlValue, XmlValue](itemPath, identity(_))
-}
-
-/**
-  *
-  * @param itemPath The XPath-like path to the element we are interested in.  (e.g. "part", "items/part", etc)
-  * @param toMappedValue Map the XmlValue into another return type, this should match the return type in MultiXmlReader
-  * @tparam XmlValue The unmarshalled element class type
-  * @tparam MappedXmlValue A mapped XmlValue type
-  */
-final case class XmlReaderPath[
-XmlValue: ClassTag,
-MappedXmlValue,
-](itemPath: String, toMappedValue: XmlValue => MappedXmlValue) extends Logging {
-  // Make this public, so we can use it for the XmlWriter JAXBContext
-  val itemClass: Class[XmlValue] = classTag[XmlValue].runtimeClass.asInstanceOf[Class[XmlValue]]
-  private[this] val jaxbContext: JAXBContext = JAXBContext.newInstance(itemClass)
-  private[this] val unmarshaller: Unmarshaller = jaxbContext.createUnmarshaller()
-
-  // The XPath-like path to the element we are interested in
-  // part => Array("part"), items/part => Array("items","part")
-  // This is reversed for easy comparison to the ArrayStack we use
-  private[xml] val path: Array[String] = itemPath.split('/')
-
-  // The name of the element we care about (last part of the path)
-  // items/part => part
-  private[xml] val itemName: String = path.last
-
-  private[xml] val targetDepth: Int = path.length - 1
-
-  private[xml] def readValue(xmlStreamReader: XMLStreamReader2): MappedXmlValue =  {
-    xmlStreamReader.require(START_ELEMENT, null, itemName)
-
-    val value: XmlValue = unmarshaller.unmarshal(xmlStreamReader, itemClass).getValue
-    toMappedValue(value)
-  }
-
-  // TODO: Cleanup the matches/pathMatches
-  // Check to see if the target depth and path match
-  private[xml] def elementMatches(xmlStreamReader: XMLStreamReader2, currentDepth: Int): Boolean = {
-    currentDepth == targetDepth && xmlStreamReader.getLocalName === itemName
-  }
-
-  private[xml] def pathMatches(xmlStreamReader: XMLStreamReader2, currentPath: mutable.ArrayStack[String]): Boolean = {
-    if (currentPath.length > targetDepth) return false
-    if (currentPath.length === targetDepth && (xmlStreamReader.getLocalName !== itemName)) return false
-
-    // Check the previous paths
-    var depth: Int = 0
-    while (depth < currentPath.length) {
-
-      // The stack indexes (in parens) change as the new paths get added
-      // items(0)                                     - current depth 1
-      // items(1) / items2(0)                         - current depth 2
-      // items(2) / items2(1) / items3 (0)            - current depth 3
-      // items(4) / items2(2) / items3 (1) / part (0) - current depth 4
-
-      // The path is static: items(0) / items2(1) / items3 (2) / part (3)
-
-      // When currentPath length is 3, the first item in the currentPath is (currentPath.length - 1), each iteration minus another 1
-      if (currentPath(currentPath.length - 1 - depth) !== path(depth)) return false
-
-      depth += 1
-    }
-
-    true
-  }
-}
-
-
 /**
   * defaultNamespaceURI - If no namespace is set, use this as a default
   * overrideDefaultNamespaceURI - Override the namespace and just use this instead
   */
-class XmlReader[T](
+final case class XmlReader[T](
   rootName: String,
   defaultNamespaceURI: String,
   overrideDefaultNamespaceURI: String,
